exceptions && validate

+++ backend/src/main/java/neu/edu/info6350/controller/ExpenseController.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 import javax.annotation.Resource;
 
+import neu.edu.info6350.exception.MyRuntimeException;
+import neu.edu.info6350.exception.PermissionException;
 import org.springframework.web.bind.annotation.*;
 
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
@@ -12,6 +14,9 @@ import neu.edu.info6350.model.Expense;
 import neu.edu.info6350.model.User;
 import neu.edu.info6350.service.ExpenseService;
 import neu.edu.info6350.service.UserService;
+
+import static neu.edu.info6350.exception.Messages.DATA_AUTH;
+import static neu.edu.info6350.exception.Messages.EXPENSE_DOES_NOT_EXIST;
 
 /**
  * @author arronshentu
@@ -24,21 +29,47 @@ public class ExpenseController {
   @Resource
   ExpenseService expenseService;
 
-  @GetMapping("/all")
+  @GetMapping("")
   List<Expense> getAll() {
     User info = userService.getInfo();
     return expenseService.list(Wrappers.<Expense>lambdaQuery().eq(Expense::getUserId, info.getId()));
   }
 
-  @DeleteMapping("/remove")
-  void remove(String id) {
+  @DeleteMapping("/{id}")
+  void remove(@PathVariable String id) {
+    if(expenseService.getById(id) == null){
+      throw new MyRuntimeException(EXPENSE_DOES_NOT_EXIST);
+    }
+    if(!expenseService.getById(id).getUserId().equals(userService.getId()) ){
+      throw new PermissionException(DATA_AUTH);
+    }
     expenseService.remove(Wrappers.<Expense>lambdaQuery().eq(Expense::getId, id));
   }
 
-  @PostMapping("/create")
-  Expense create(Expense expense) {
+  @PostMapping("")
+  Expense create(@RequestBody Expense expense) {
     expense.setUserId(userService.getId());
     expenseService.save(expense);
+    return expense;
+  }
+
+  @GetMapping("/{id}")
+  Expense get(@PathVariable String id) {
+    if(expenseService.getById(id) == null){
+      throw new MyRuntimeException(EXPENSE_DOES_NOT_EXIST);
+    }
+    if(!expenseService.getById(id).getUserId().equals(userService.getId()) ){
+      throw new PermissionException(DATA_AUTH);
+    }
+    return expenseService.getById(id);
+  }
+
+  @PutMapping("")
+  Expense update(@RequestBody Expense expense) {
+    if(!expense.getUserId().equals(userService.getId())) {
+      throw new RuntimeException(DATA_AUTH);
+    }
+    expenseService.updateById(expense);
     return expense;
   }
